requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.2
alembic==1.13.3
pydantic==2.9.2
pydantic[email]==2.5.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0 
passlib[bcrypt]==1.7.4  
python-multipart==0.0.12
pillow==10.4.0 
python-magic==0.4.27  
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
celery